fix(services-overview) renamed to ServicesCards
