+++ components/services-cards/services-cards.less
@@ -1,6 +1,6 @@
 @import (reference) '../../styles/variables.less';
 
-.services-overview {
+.services-cards {
 	position: relative;
 	width: 100%;
 	height: 100%;
