+++ styles/index.less
@@ -64,7 +64,7 @@
 @import (once) '../components/rich-body-text/rich-body-text';
 @import (once) '../components/select-dropdown/select-dropdown';
 @import (once) '../components/service-card/service-card';
-@import (once) '../components/services-overview/services-overview';
+@import (once) '../components/services-cards/services-cards';
 @import (once) '../components/services-overview-page/services-overview-page.less';
 @import (once) '../components/shapes/shapes';
 @import (once) '../components/social-media/social-media';
